actors are now displayed and select-able
