+++ src/pages/authentication/RegisterPage/RegisterPage.module.css
@@ -9,11 +9,9 @@
 .register-button { composes: gameboy-start-button mb-3 from global; }
 .already-have-account { composes: m-1 from global; font-size: 0.7em; }
 .figures-container {
-    display: grid;
-    grid-template-columns: 60px 60px 60px;
-}
-
-.figures-container * {
+    margin-top: 32px;
+    margin-bottom: 32px;
     width: 100%;
-    height: 60px;
+    display: grid;
+    grid-template-columns: 96px 96px 96px;
 }
